\name{d}
\alias{d}
\docType{data}
\title{
Example Dataset for tab Package
}
\description{
This dataset is used to illustrate the various functions in the R package tab.
}
\usage{data(d)}
\format{
  A data frame with 300 observations on the following 15 variables.
  \describe{
    \item{\code{ID}}{a numeric vector}
    \item{\code{Group}}{a factor}
    \item{\code{Age}}{a numeric vector}
    \item{\code{Sex}}{a factor}
    \item{\code{Race}}{a factor}
    \item{\code{BMI}}{a numeric vector}
    \item{\code{time}}{a numeric vector}
    \item{\code{delta}}{a numeric vector}
    \item{\code{death_1yr}}{a numeric vector}
    \item{\code{bp.1}}{a numeric vector}
    \item{\code{bp.2}}{a numeric vector}
    \item{\code{bp.3}}{a numeric vector}
    \item{\code{highbp.1}}{a numeric vector}
    \item{\code{highbp.2}}{a numeric vector}
    \item{\code{highbp.3}}{a numeric vector}
  }
}
\details{
NA
}
\source{
This dataset was generated in R.
}
\references{
Acknowledgment: This material is based upon work supported by the National Science Foundation Graduate Research Fellowship under Grant No. DGE-0940903.
}
\examples{
NA
}
\keyword{datasets}