% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_v.R
\name{glm_v}
\alias{glm_v}
\title{Print a GLM Summary Table to the RStudio Viewer}
\usage{
glm_v(...)
}
\arguments{
\item{...}{Arguments to pass to glm.}
}
\value{
kable
}
\description{
You can call this function as you would \code{\link[stats]{glm}} or pass a
previously fitted \code{\link[stats]{glm}} object. Either way, the result is
a summary table printed to the Viewer.
}
\examples{
# Fit and view
glm_v(death_1yr ~ Age + Sex + Race, data = tabdata, family = "binomial")

# Fit then view
fit <- glm(death_1yr ~ Age + Sex + Race, data = tabdata, family = "binomial")
glm_v(fit)

# Piping is OMG so cool Hashtag HexStickerz
fit \%>\% glm_v()

}
