% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hide_rows.R
\name{hide_rows}
\alias{hide_rows}
\title{Hide rows of selected levels of dichotomous variables from a Table
(e.g. "no" category of Smoking)}
\usage{
hide_rows(tab, rows, ...)
}
\arguments{
\item{tab}{Data frame.}

\item{rows}{Character string with name of levels (rows) which should be hidden.}
}
\value{
Data frame which you can print in R (e.g. with \strong{xtable}'s
\code{\link[xtable]{xtable}} or \strong{knitr}'s \code{\link[knitr]{kable}}).
}
\description{
Hide rows of selected levels of dichotomous variables from a Table
(e.g. "no" category of Smoking)
}
\examples{
# Compare age, sex, and race by treatment group
# Hide level "Male" of Variable "sex"
tabmulti(Age + Sex + Race ~ Group, data = tabdata) \%>\% hide_rows(rows = c("male")) \%>\% kable()


}
