% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\title{Crosstabulation of survey data (Weighted or unweighted data)}
\usage{
tab(var1, var2 = NULL, weight = NULL, col = FALSE, row = FALSE, cell = FALSE)
}
\arguments{
\item{var1}{Input Variable one for survey data}

\item{var2}{Input Variable Two for survey data. By default is null when a single crosstabulation is needed}

\item{weight}{Attach the weight for data if  you want to weight them. By default is null when weight is not needed}

\item{col}{Is used when the crosstabulation is needed on column, by default is False make it true when the crosstabulation is needed by column}

\item{row}{Is used when the crosstabulation is needed on row, by default is False make it true when the crosstabulation is needed by row}

\item{cell}{Is used when the crosstabulation is needed on cell, by default is False make it true when the crosstabulation is needed by cell}
}
\description{
This function helps the crosstabulation of Single, Two variables either weighted with survey weight or unweighted.
It is based on tabulate function like the one used in Stata,it is helpful for survey data.
}
\examples{
tab()
}
\keyword{tab}
