% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabcox_old.R
\name{tabcox_old}
\alias{tabcox_old}
\title{Generate Summary Table for Fitted Cox Proportional Hazards (Cox PH) Model}
\usage{
tabcox_old(x, time, delta, latex = FALSE, xlabels = NULL, cluster = NULL,
  robust = TRUE, decimals = 2, p.decimals = c(2, 3), p.cuts = 0.01,
  p.lowerbound = 0.001, p.leading0 = TRUE, p.avoid1 = FALSE, n = FALSE,
  events = FALSE, stand.x = FALSE, greek.beta = FALSE,
  binary.compress = TRUE, bold.colnames = TRUE, bold.varnames = FALSE,
  bold.varlevels = FALSE, variable.colname = "Variable",
  suppress.beta = FALSE, print.html = FALSE,
  html.filename = "table1.html")
}
\arguments{
\item{x}{For single predictor, vector of values; for multiple predictors,
data frame with one column per predictor. Categorical variables should be of
class "factor."}

\item{time}{Numeric vector with times to event or censoring.}

\item{delta}{Numeric vector where 1 = event observed, 0 = censored.}

\item{latex}{Logical value for whether to format table for printing in LaTeX,
e.g. using \code{\link[xtable]{xtable}} [1].}

\item{xlabels}{Character vector with labels for the \code{x} variables and
their levels. Often useful to leave as \code{NULL} first, see how the table
gets displayed, and then re-run with meaningful labels where they need to be.}

\item{cluster}{Vector indicating clusters of subjects.}

\item{robust}{Logical value for whether robust sandwich method standard
errors should be used. Only used if \code{cluster} is specified.}

\item{decimals}{Numeric value specifying number of decimal places for numbers
other than p-values. If unspecified, function tries to make a reasonable
choice based on the data.}

\item{p.decimals}{Numeric value specifying number of decimal places for
p-values. Can be a vector if you want the number of decimals to depend on
what range the p-value is in. See \code{p.cuts}.}

\item{p.cuts}{Numeric value or vector of cutpoints to control number of
decimal places for p-values. For example, by default \code{p.cuts = 0.1} and
\code{p.decimals = c(2, 3)}, meaning p-values in the range [0.1, 1] are
printed to 2 decimal places while p-values in the range [0, 0.1) are printed
to 3.}

\item{p.lowerbound}{Numeric value specifying cutpoint beneath which p-values
appear as <p.lowerbound.}

\item{p.leading0}{Logical value for whether p-values should appear with
leading 0's before the decimal point.}

\item{p.avoid1}{Logical value for whether p-values that round to 1 should
appear as \code{>0.99} (or similarly depending on \code{p.decimals} and
\code{p.cuts}) rather than 1.}

\item{n}{Logical value for whether to include a sample size column.}

\item{events}{Logical value for whether to include a number of events
(i.e. uncensored observations) column.}

\item{stand.x}{Logical value for whether to use standardized regression
coefficients for continuous predictors.}

\item{greek.beta}{Logical value for whether to display the Greek letter beta
rather than "Beta". Only used if \code{latex = TRUE}.}

\item{binary.compress}{Logical value for whether binary \code{x} factor
variables should be displayed as a single row rather than two.}

\item{bold.colnames}{Logical value for whether to use bold font for column
headings. Only used if \code{latex = TRUE}.}

\item{bold.varnames}{Logic value for whether to use bold font for the
\code{y} variable name in the first column. Only used if \code{latex = TRUE}.}

\item{bold.varlevels}{Logical value for whether levels of factor \code{x}
variables should appear in bold font in the first column. Only used if
\code{latex = TRUE}.}

\item{variable.colname}{Character string with desired heading for first
column of table, in case you prefer something other than \code{"Variable"}.}

\item{suppress.beta}{Logical value for whether to suppress Beta (SE) column.}

\item{print.html}{Logical value for whether to write a .html file with the
table to the current working directory.}

\item{html.filename}{Character string specifying the name of the .html file
that gets written if \code{print.html = TRUE}.}
}
\value{
Character matrix with table summarizing the fitted Cox PH model.
}
\description{
Fits a Cox PH regression model using the \pkg{survival} package [1, 2] and
creates a table summarizing the results.
}
\section{Note}{

If you want to paste your tables into Microsoft Word, you can use either of
these approaches:

\enumerate{

\item 1. Use the \code{\link[Kmisc]{write.cb}} function in \pkg{Kmisc} [2].
If your table is stored in a character matrix named \code{table1}, use
\code{write.cb(table1)} to copy the table to your clipboard. Paste the result
into your document, then highlight the text and go to
\code{Insert -> Table -> Convert Text to Table... OK}.

\item 2. Set \code{print.html = TRUE}. This will result in a .html file
being written to your current working directory. When you open this file, you
will see a (hopefully) nice-looking table that you can copy and paste into
your document. You can control the name of this file with
\code{html.filename}.
}

If you want to use LaTeX, R Markdown, knitr, Sweave, etc., set
\code{latex = TRUE} and then use \code{\link[xtable]{xtable}} [1]. You may
have to set \code{sanitize.text.function = identity} when calling
\code{\link[xtable]{print.xtable}}.
}

\examples{
# Load in sample dataset d and drop rows with missing values
data(d)
d <- d[complete.cases(d), ]

# Create labels for race levels
races <- c("White", "Black", "Mexican American", "Other")

# Test whether race is associated with survival
coxtable1 <- tabcox_old(x = d$Race, time = d$time, delta = d$delta,
                        xlabels = c("Race", races))

# Test whether age, sex, race, and treatment group are associated with
# survival
coxtable2 <-
  tabcox_old(x = d[,c("Age", "Sex", "Race", "Group")],
             time = d$time, delta = d$delta,
             xlabels = c("Age", "Male", "Race", races, "Treatment"))


}
\references{
1. Therneau, T. (2015). A Package for Survival Analysis in S. R package
version 2.38. \url{https://cran.r-project.org/package=survival}.

2. Therneau, T.M. and Grambsch, P.M. (2000). Modeling Survival Data:
Extending the Cox Model. Springer, New York. ISBN 0-387-98784-3.

3. Dahl, D.B. (2016). xtable: Export Tables to LaTeX or HTML. R package
version 1.8-2, \url{https://cran.r-project.org/package=xtable}.

4. Ushley, K. (2013). Kmisc: Kevin Miscellaneous. R package version 0.5.0.
\url{https://CRAN.R-project.org/package=Kmisc}.
}
