% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab-data.R
\docType{data}
\name{tabsvydata}
\alias{tabsvydata}
\title{Sample Survey Dataset for \pkg{tab} Package}
\source{
https://wwwn.cdc.gov/Nchs/Nhanes/2003-2004/DEMO_C.htm
}
\description{
Data frame with with 9 variables, used to illustrate certain functions. Data
are derived from the National Health and Nutrition Examination Survey, years
2003-2004, although the variables 'time' and 'event' are simulated (fake).
}
\references{
Centers for Disease Control and Prevention (CDC). National Center for Health
Statistics (NCHS). National Health and Nutrition Examination Survey Data.
Hyattsville, MD: US Department of Health and Human Services, Centers for
Disease Control and Prevention, 2003-2004.
https://wwwn.cdc.gov/nchs/nhanes/continuousnhanes/default.aspx?BeginYear=2003.
Accessed June 8, 2019.
}
