% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toviewer.R
\name{toviewer}
\alias{toviewer}
\title{Output a Table to the RStudio Viewer}
\usage{
toviewer(x)
}
\arguments{
\item{x}{Character matrix}
}
\description{
Does some basic formatting and then calls \code{\link[knitr]{kable}} and
\code{\link[kableExtra]{kable_styling}} to print table to Viewer.
}
