\name{d}
\alias{d}
\docType{data}
\title{
Example dataset for R package tab. 
}
\description{
This dataset is used to illustrate the various functions in the R package tab.
}
\usage{data(d)}
\format{
  A data frame with 245 observations on the following 9 variables.
  \describe{
    \item{\code{id}}{a numeric vector}
    \item{\code{group}}{a numeric vector}
    \item{\code{age}}{a numeric vector}
    \item{\code{sex}}{a numeric vector}
    \item{\code{race}}{a factor with levels \code{1} \code{2} \code{3} \code{99}}
    \item{\code{bmi}}{a numeric vector}
    \item{\code{time}}{a numeric vector}
    \item{\code{delta}}{a numeric vector}
    \item{\code{death_1yr}}{a numeric vector}
  }
}
\details{
NA
}
\source{
This dataset was generated in R.
}
\references{
NA
}
\examples{
NA
}
\keyword{datasets}