% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabmeans.R
\name{tabmeans}
\alias{tabmeans}
\title{Generate Summary Table Comparing Group Means}
\usage{
tabmeans(x, y, latex = FALSE, variance = "unequal", xname = NULL,
  xlevels = NULL, yname = NULL, quantiles = NULL, quantile.vals = FALSE,
  parenth = "sd", text.label = NULL, parenth.sep = "-", decimals = NULL,
  p.include = TRUE, p.decimals = c(2, 3), p.cuts = 0.01,
  p.lowerbound = 0.001, p.leading0 = TRUE, p.avoid1 = FALSE,
  overall.column = TRUE, n.column = FALSE, n.headings = TRUE,
  bold.colnames = TRUE, bold.varnames = FALSE,
  variable.colname = "Variable", fig = FALSE, print.html = FALSE,
  html.filename = "table1.html", ...)
}
\arguments{
\item{x}{Vector of values for the categorical \code{x} variable.}

\item{y}{Vector of values for the continuous \code{y} variable.}

\item{latex}{Logical value for whether to format table for printing in LaTeX,
e.g. using \code{\link[xtable]{xtable}} [1].}

\item{variance}{Character string specifying which version of the two-sample
t-test to use, supposing \code{x} has two levels. Choices are \code{"equal"}
for equal variance t-test, \code{"unequal"} for unequal variance t-test, and
\code{"ftest"} for F test to determine which to use.}

\item{xname}{Character string with a label for the \code{x} variable.}

\item{xlevels}{Character vector with labels for each level of \code{x}, for
column headings.}

\item{yname}{Character string with a label for the \code{y} variable.}

\item{quantiles}{Numeric value. If specified, table compares mean \code{y}
across quantiles of \code{x}. For example, if \code{x} contains BMI values
and \code{y} HDL levels, setting \code{quantiles = 3} compares mean HDL
across BMI tertiles.}

\item{quantile.vals}{Logical value for whether labels for \code{x} should
show quantile number and corresponding range, e.g. Q1 [0.00, 0.25), rather
than just the quantile number.}

\item{parenth}{Character string specifying what statistic should appear in
parentheses after the means in each cell. Choices are \code{"none"},
\code{"sd"} for standard deviation, \code{"se"} for standard error,
\code{"t.ci"} for 95\% confidence interval based on t distribution, and
\code{"z.ci"} for 95\% confidence interval based on z distribution.}

\item{text.label}{Character string with text to put after the \code{y}
variable name, identifying what cell values and parentheses represent.}

\item{parenth.sep}{Character string with separator to place between first and
second numbers in parentheses, if applicable. Typically \code{"-"} or
\code{", "}.}

\item{decimals}{Numeric value specifying number of decimal places for numbers
other than p-values. If unspecified, function tries to make a reasonable
choice based on the data.}

\item{p.include}{Logical value for whether to include a p-value column.}

\item{p.decimals}{Numeric value specifying number of decimal places for
p-values. Can be a vector if you want the number of decimals to depend on
what range the p-value is in. See \code{p.cuts}.}

\item{p.cuts}{Numeric value or vector of cutpoints to control number of
decimal places for p-values. For example, by default \code{p.cuts = 0.1} and
\code{p.decimals = c(2, 3)}, meaning p-values in the range [0.1, 1] are
printed to 2 decimal places while p-values in the range [0, 0.1) are printed
to 3.}

\item{p.lowerbound}{Numeric value specifying cutpoint beneath which p-values
appear as <p.lowerbound.}

\item{p.leading0}{Logical value for whether p-values should appear with
leading 0's before the decimal point.}

\item{p.avoid1}{Logical value for whether p-values that round to 1 should
appear as \code{>0.99} (or similarly depending on \code{p.decimals} and
\code{p.cuts}) rather than 1.}

\item{overall.column}{Logical value for whether to include a column with the
overall mean of \code{y}.}

\item{n.column}{Logical value for whether to include a column with the
overall sample size.}

\item{n.headings}{Logical value for whether to include sample sizes in
parentheses in column headings.}

\item{bold.colnames}{Logical value for whether to use bold font for column
headings. Only used if \code{latex = TRUE}.}

\item{bold.varnames}{Logic value for whether to use bold font for the
\code{y} variable name in the first column. Only used if \code{latex = TRUE}.}

\item{variable.colname}{Character string with desired heading for first
column of table, in case you prefer something other than \code{"Variable"}.}

\item{fig}{Logical value for whether to generate a figure rather than a
table.}

\item{print.html}{Logical value for whether to write a .html file with the
table to the current working directory.}

\item{html.filename}{Character string specifying the name of the .html file
that gets written if \code{print.html = TRUE}.}

\item{...}{Additional arguments to pass to \code{\link[graphics]{plot}}. Only
used if \code{fig = TRUE}.}
}
\value{
Character matrix with table comparing mean \code{y} across levels of
\code{x}.
}
\description{
Creates table (or figure) comparing the mean of \code{y} across levels of
\code{x}.
}
\details{
A t-test is used to compare means if \code{x} has two levels, and a one-way
analysis of variance is used if \code{x} has more than two levels.
Observations with missing values for \code{x} and/or \code{y} are dropped.
}
\section{Note}{

If you want to paste your tables into Microsoft Word, you can use either of
these approaches:

\enumerate{

\item 1. Use the \code{\link[Kmisc]{write.cb}} function in \pkg{Kmisc} [2].
If your table is stored in a character matrix named \code{table1}, use
\code{write.cb(table1)} to copy the table to your clipboard. Paste the result
into your document, then highlight the text and go to
\code{Insert -> Table -> Convert Text to Table... OK}.

\item 2. Set \code{print.html = TRUE}. This will result in a .html file
being written to your current working directory. When you open this file, you
will see a (hopefully) nice-looking table that you can copy and paste into
your document. You can control the name of this file with
\code{html.filename}.
}

If you want to use LaTeX, R Markdown, knitr, Sweave, etc., set
\code{latex = TRUE} and then use \code{\link[xtable]{xtable}} [1]. You may
have to set \code{sanitize.text.function = identity} when calling
\code{\link[xtable]{print.xtable}}.
}

\examples{
# Load in sample dataset d and drop rows with missing values
data(d)
d <- d[complete.cases(d), ]

# Compare mean BMI in control group vs. treatment group - table and figure
meanstable1 <- tabmeans(x = d$Group, y = d$BMI)
meansfig1 <- tabmeans(x = d$Group, y = d$BMI, fig = TRUE)

# Compare mean BMI by race - table and figure
meanstable2 <- tabmeans(x = d$Race, y = d$BMI)
meansfig2 <- tabmeans(x = d$Race, y = d$BMI, fig = TRUE)

# Compare mean baseline systolic BP across tertiles of BMI - table and figure
meanstable3 <- tabmeans(x = d$BMI, y = d$bp.1, yname = "Systolic BP",
                        quantiles = 3)
meansfig3 <- tabmeans(x = d$BMI, y = d$bp.1, quantiles = 3, fig = TRUE,
                      yname = "Systolic BP", xname = "BMI Tertile")

# Create single table comparing mean BMI and mean age in control vs.
# treatment group
meanstable4 <- rbind(tabmeans(x = d$Group, y = d$BMI),
                     tabmeans(x = d$Group, y = d$Age))

# An easier way to make the above table is to call the tabmulti function
meanstable5 <- tabmulti(dataset = d, xvarname = "Group",
                        yvarnames = c("BMI", "Age"))

# meanstable4 and meanstable5 are equivalent
all(meanstable4 == meanstable5)


}
\references{
1. Dahl, D.B. (2016). xtable: Export Tables to LaTeX or HTML. R package
version 1.8-2, \url{https://cran.r-project.org/package=xtable}.

2. Ushley, K. (2013). Kmisc: Kevin Miscellaneous. R package version 0.5.0.
\url{https://CRAN.R-project.org/package=Kmisc}.
}
