% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabmedians.R
\name{tabmedians}
\alias{tabmedians}
\title{Generate Summary Table Comparing Group Medians}
\usage{
tabmedians(x, y, columns = c("xgroups", "p"), parenth = "iqr",
  sep.char = ", ", xname = NULL, xlevels = NULL, yname = NULL,
  text.label = NULL, quantiles = NULL, quantile.vals = FALSE,
  decimals = NULL, p.decimals = c(2, 3), p.cuts = 0.01,
  p.lowerbound = 0.001, p.leading0 = TRUE, p.avoid1 = FALSE,
  n.headings = TRUE, variable.colname = "Variable", fig = FALSE,
  print.html = FALSE, html.filename = "table1.html", ...)
}
\arguments{
\item{x}{Vector of values for the categorical \code{x} variable.}

\item{y}{Vector of values for the continuous \code{y} variable.}

\item{columns}{Character vector specifying what columns to include. Choices
for each element are \code{"n"} for total sample size, \code{"overall"} for
overall median, \code{"xgroups"} for \code{x} group medians, \code{"diff"}
for difference in \code{x} group medians (only available for binary
\code{x}), \code{"test"} for test statistic, and \code{"p"} for p-value.}

\item{parenth}{Character string specifying what values are shown in
parentheses after the medians in each cell. Choices are \code{"none"},
\code{"iqr"}, \code{"q1q3"} for first and third quartiles, \code{"range"},
\code{"minmax"}, and \code{"ci"} for 95\% confidence interval for the medians
based on normal approximation to binomial.}

\item{sep.char}{Character string with separator to place between lower and
upper bound of confidence intervals. Typically \code{"-"} or \code{", "}.}

\item{xname}{Character string with a label for the \code{x} variable.}

\item{xlevels}{Character vector with labels for the levels of \code{x}, used
in column headings.}

\item{yname}{Character string with a label for the \code{y} variable.}

\item{text.label}{Character string with text to put after the \code{y}
variable name, identifying what cell values and parentheses represent.}

\item{quantiles}{Numeric value. If specified, table compares \code{y} across
quantiles of \code{x} created on the fly.}

\item{quantile.vals}{Logical value for whether labels for \code{x} quantiles
should show quantile number and corresponding range, e.g. Q1 [0.00, 0.25),
rather than just the quantile number.}

\item{decimals}{Numeric value specifying number of decimal places for numbers
other than p-values.}

\item{p.decimals}{Numeric value specifying number of decimal places for
p-values. Can be a vector if you want the number of decimals to depend on
what range the p-value is in. See \code{p.cuts}.}

\item{p.cuts}{Numeric value or vector of cutpoints to control number of
decimal places for p-values. For example, by default \code{p.cuts = 0.1} and
\code{p.decimals = c(2, 3)}, meaning p-values in the range [0.1, 1] are
printed to 2 decimal places while p-values in the range [0, 0.1) are printed
to 3.}

\item{p.lowerbound}{Numeric value specifying cutpoint beneath which p-values
appear as <p.lowerbound.}

\item{p.leading0}{Logical value for whether p-values should appear with
leading 0's before the decimal point.}

\item{p.avoid1}{Logical value for whether p-values that round to 1 should
appear as \code{>0.99} (or similarly depending on \code{p.decimals} and
\code{p.cuts}) rather than 1.}

\item{n.headings}{Logical value for whether to display group sample sizes in
parentheses in column headings.}

\item{variable.colname}{Character string with desired heading for first
column of table, in case you prefer something other than \code{"Variable"}.}

\item{fig}{Logical value for whether to generate a figure rather than a
table.}

\item{print.html}{Logical value for whether to write a .html file with the
table to the current working directory.}

\item{html.filename}{Character string specifying the name of the .html file
that gets written if \code{print.html = TRUE}.}

\item{...}{Additional arguments to pass to \code{\link[graphics]{plot}}. Only
used if \code{fig = TRUE}.}
}
\value{
Character matrix comparing median \code{y} across levels of \code{x}.
}
\description{
Creates table (or figure) comparing the median of \code{y} across levels of
\code{x}.
}
\details{
If \code{x} has 2 levels, a Mann-Whitney U (also known as Wilcoxon
rank-sum) test is used to test whether the distribution of \code{y} differs
in the two groups; if \code{x} has more than 2 levels, a Kruskal-Wallis test
is used to test whether the distribution of \code{y} differs across at
least two of the groups. Observations with missing values for \code{x} and/or
\code{y} are dropped.
}
\section{Note}{

If you want to paste your tables into Microsoft Word, you can use either of
these approaches:

\enumerate{

\item Use the \code{\link[Kmisc]{write.cb}} function in \pkg{Kmisc} [2].
If your table is stored in a character matrix named \code{table1}, use
\code{write.cb(table1)} to copy the table to your clipboard. Paste the result
into your document, then highlight the text and go to
\code{Insert -> Table -> Convert Text to Table... OK}.

\item Set \code{print.html = TRUE}. This will result in a .html file
being written to your current working directory. When you open this file, you
will see a (hopefully) nice-looking table that you can copy and paste into
your document. You can control the name of this file with
\code{html.filename}.
}

If you want to use LaTeX, R Markdown, knitr, Sweave, etc., set
\code{latex = TRUE}. Then, you can pass the returned object to
\code{\link[xtable]{xtable}} or \code{\link[knitr]{kable}} for printing. For
\code{\link[xtable]{xtable}}, you may have to set
\code{sanitize.text.function = identity} when printing.
}

\examples{
# Load in sample dataset and drop rows with missing values
data(tabdata)
tabdata <- tabdata[complete.cases(tabdata), ]

# Compare median BMI in control group vs. treatment group
(medtable1 <- tabmedians(x = tabdata$Group, y = tabdata$BMI))

# Compare median BMI by quartile of age
(medtable2 <- tabmedians(x = tabdata$Age, y = tabdata$BMI, quantiles = 4))

# Create single table comparing median BMI and median age in control vs.
# treatment group
(medtable3 <- rbind(tabmedians(x = tabdata$Group, y = tabdata$BMI),
                    tabmedians(x = tabdata$Group, y = tabdata$Age)))

# An easier way to make this table is to use tabmulti
medtable4 <- tabmulti(data = tabdata, xvarname = "Group",
                      yvarnames = c("BMI", "Age"), ymeasures = "median")


}
\references{
1. Dahl, D.B. (2016). xtable: Export Tables to LaTeX or HTML. R package
version 1.8-2, \url{https://cran.r-project.org/package=xtable}.

2. Ushley, K. (2013). Kmisc: Kevin Miscellaneous. R package version 0.5.0.
\url{https://CRAN.R-project.org/package=Kmisc}.
}
